+++ bitbybit-tests/src/bitfield_tests.rs
@@ -1819,6 +1819,32 @@ fn test_debug_impl() {
 }
 
 #[test]
+fn test_debug_write_only_fields_skipped() {
+    #[bitfield(u16, debug)]
+    struct Test {
+        #[bit(15, w)]
+        weird_control_bit: bool,
+
+        #[bits(0..=7, rw)]
+        lower: u8,
+    }
+    let test = Test::new_with_raw_value(0x1F2F);
+    let display_str = format!("{:?}", test);
+    assert_eq!(display_str, "Test { lower: 47 }");
+}
+#[test]
+fn test_debug_array_support() {
+    #[bitbybit::bitfield(u8, debug)]
+    struct ArrayField {
+        #[bits(0..=1, rw)]
+        fields: [u2; 4],
+    }
+    let test = ArrayField::new_with_raw_value(0b00011011);
+    let display_str = format!("{:?}", test);
+    assert_eq!(display_str, "ArrayField { fields: [3, 2, 1, 0] }");
+}
+
+#[test]
 fn introspection() {
     // Put into module to ensure everything's pub
     mod test {